properly incorporating imgui library
