+++ src/PoeFilterParser/packages.config
@@ -1,6 +1,6 @@
 ﻿<?xml version="1.0" encoding="utf-8"?>
 <packages>
-  <package id="Antlr4" version="4.6.4" targetFramework="net462" developmentDependency="true" />
-  <package id="Antlr4.CodeGenerator" version="4.6.4" targetFramework="net462" developmentDependency="true" />
-  <package id="Antlr4.Runtime" version="4.6.4" targetFramework="net462" />
+  <package id="Antlr4" version="4.6.4" targetFramework="net46" developmentDependency="true" />
+  <package id="Antlr4.CodeGenerator" version="4.6.4" targetFramework="net46" developmentDependency="true" />
+  <package id="Antlr4.Runtime" version="4.6.4" targetFramework="net46" />
 </packages>
