+++ src/app.config
@@ -2,5 +2,5 @@
 <configuration>
   <startup>
     
-  <supportedRuntime version="v4.0" sku=".NETFramework,Version=v4.6.2"/></startup>
+  <supportedRuntime version="v4.0" sku=".NETFramework,Version=v4.6"/></startup>
 </configuration>
