+++ src/packages.config
@@ -1,13 +1,14 @@
 ﻿<?xml version="1.0" encoding="utf-8"?>
 <packages>
-  <package id="Antlr4.Runtime" version="4.6.4" targetFramework="net462" />
-  <package id="Newtonsoft.Json" version="10.0.3" targetFramework="net462" />
-  <package id="SharpDX" version="4.0.1" targetFramework="net462" />
-  <package id="SharpDX.Desktop" version="4.0.1" targetFramework="net462" />
-  <package id="SharpDX.Direct3D9" version="4.0.1" targetFramework="net462" />
-  <package id="SharpDX.Mathematics" version="4.0.1" targetFramework="net462" />
-  <package id="System.Buffers" version="4.3.0" targetFramework="net462" />
-  <package id="System.Numerics.Vectors" version="4.4.0" targetFramework="net462" />
-  <package id="System.Runtime.CompilerServices.Unsafe" version="4.4.0" targetFramework="net462" />
-  <package id="Trinet.Core.IO.Ntfs" version="4.0.0" targetFramework="net462" />
+  <package id="Antlr4.Runtime" version="4.6.4" targetFramework="net46" />
+  <package id="ImGui.NET" version="0.4.0" targetFramework="net46" />
+  <package id="Newtonsoft.Json" version="10.0.3" targetFramework="net46" />
+  <package id="SharpDX" version="4.0.1" targetFramework="net46" />
+  <package id="SharpDX.Desktop" version="4.0.1" targetFramework="net46" />
+  <package id="SharpDX.Direct3D9" version="4.0.1" targetFramework="net46" />
+  <package id="SharpDX.Mathematics" version="4.0.1" targetFramework="net46" />
+  <package id="System.Buffers" version="4.4.0" targetFramework="net46" />
+  <package id="System.Numerics.Vectors" version="4.4.0" targetFramework="net46" />
+  <package id="System.Runtime.CompilerServices.Unsafe" version="4.4.0" targetFramework="net46" requireReinstallation="true" />
+  <package id="Trinet.Core.IO.Ntfs" version="4.0.0" targetFramework="net46" />
 </packages>
